silk/fabric_tools/fabric_contrib_api.py
from fabric.contrib.project import rsync_project, upload_project
from fabric.contrib.console import confirm
from fabric.contrib.files import exists, is_link, first, upload_template, sed, uncomment, comment, contains, append

__all__ = ["rsync_project",
           "upload_project",
           "confirm",
           "exists",
           "is_link",
           "first",
           "upload_template",
           "sed",
           "uncomment",
           "comment",
           "contains",
           "append"]
