requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.3.1
Fabric3==1.14.post1
idna==2.7
more-itertools==4.3.0
paramiko==2.4.1
pathlib2==2.3.2
pluggy==0.7.1
py==1.6.0
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
pytest==3.7.4
pytest-cov==2.6.0
requests==2.19.1
six==1.11.0
urllib3==1.23
